+++ src/modules/systemlib/system_params.c
@@ -38,7 +38,7 @@
  */
 
 #include <px4_config.h>
-#include <systemlib/param/param.h>
+#include <param/param.h>
 
 /**
  * Auto-start script index.
